Add signup and login functionality

+++ springapp/simplelibrary/src/main/java/org/simplelibrary/repository/AuthGroupRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthGroupRepository extends JpaRepository<AuthGroup, Long> {
+public interface AuthGroupRepository extends JpaRepository<AuthGroup, Integer> {
 
 }
